backend/pyproject.toml
[tool.poetry]
name = "moto-backend"
version = "1.0.0"
description = "MOTO System - Django REST API"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"

# Django core
Django = "~5.2"
djangorestframework = "~3.16"

# Database
psycopg = {version = "~3.3", extras = ["binary"]}

# Cache
django-redis = "~5.4"

# Django ecosystem
django-cors-headers = "~4.9"
django-filter = ">=25.2"
django-environ = "~0.12"
django-model-utils = "~5.0"
drf-spectacular = "~0.29"

# WSGI server
gunicorn = ">=25.0"

# Data validation
pydantic = "~2.12"

[tool.poetry.group.dev.dependencies]
# Code quality
ruff = ">=0.15"
mypy = "~1.19"
django-stubs = {version = "~5.2", extras = ["compatible-mypy"]}
djangorestframework-stubs = {version = "~3.16", extras = ["compatible-mypy"]}

[tool.poetry.group.test.dependencies]
pytest = "~9.0"
pytest-django = "~4.11"
pytest-cov = "~7.0"
freezegun = "~1.5"
factory-boy = "~3.3"
Faker = ">=40.0"
pytest-postgresql = "~8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ============================================================================
# Tool Configuration (all in one file!)
# ============================================================================

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-v",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
    "e2e: marks end-to-end tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/management/commands/*",
    "manage.py",
    "config/*",
]

[tool.coverage.report]
fail_under = 85
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["."]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "S",     # flake8-bandit (security)
    "A",     # flake8-builtins
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "DJ",    # flake8-django
    "T20",   # flake8-print
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific rules
]
ignore = [
    "S101",  # assert used (needed for pytest)
    "S104",  # binding to all interfaces (needed for Docker)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "S106"]
"config/**/*" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.mypy]
python_version = "3.13"
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "django_filters.*",
    "environ",
    "environ.*",
    "factory.*",
    "faker.*",
    "freezegun.*",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"
